// src/types.ts
export type ResumeUpdater =
  | ResumeUpdater.UpdateContactIcons
  | ResumeUpdater.UpdateName
  | ResumeUpdater.UpdateSections
  | ResumeUpdater.UpdateSettings
  | ResumeUpdater.UpdateSummary;

export namespace ResumeUpdater {
  export enum Type {
    UpdateContactIcons = "me.kthomas.resume.ResumeUpdater.UpdateContactIcons",
    UpdateName = "me.kthomas.resume.ResumeUpdater.UpdateName",
    UpdateSections = "me.kthomas.resume.ResumeUpdater.UpdateSections",
    UpdateSettings = "me.kthomas.resume.ResumeUpdater.UpdateSettings",
    UpdateSummary = "me.kthomas.resume.ResumeUpdater.UpdateSummary",
  }
  
  export interface UpdateContactIcons {
    type: ResumeUpdater.Type.UpdateContactIcons;
    updater: ModifyContactItems;
  }
  
  export interface UpdateName {
    type: ResumeUpdater.Type.UpdateName;
    newName: string;
  }
  
  export interface UpdateSections {
    type: ResumeUpdater.Type.UpdateSections;
    updater: ModifySections;
  }
  
  export interface UpdateSettings {
    type: ResumeUpdater.Type.UpdateSettings;
    updater: SettingsUpdater;
  }
  
  export interface UpdateSummary {
    type: ResumeUpdater.Type.UpdateSummary;
    newSummary: string;
  }
}

export type ModifyContactItems =
  | ModifyContactItems.AddContactItem
  | ModifyContactItems.MoveContactItem
  | ModifyContactItems.RemoveContactItem
  | ModifyContactItems.UpdateContactItem;

export namespace ModifyContactItems {
  export enum Type {
    AddContactItem = "me.kthomas.resume.ModifyContactItems.AddContactItem",
    MoveContactItem = "me.kthomas.resume.ModifyContactItems.MoveContactItem",
    RemoveContactItem = "me.kthomas.resume.ModifyContactItems.RemoveContactItem",
    UpdateContactItem = "me.kthomas.resume.ModifyContactItems.UpdateContactItem",
  }
  
  export interface AddContactItem {
    type: ModifyContactItems.Type.AddContactItem;
    item: ResumeContactItem;
  }
  
  export interface MoveContactItem {
    type: ModifyContactItems.Type.MoveContactItem;
    oldIndex: number;
    newIndex: number;
  }
  
  export interface RemoveContactItem {
    type: ModifyContactItems.Type.RemoveContactItem;
    index: number;
  }
  
  export interface UpdateContactItem {
    type: ModifyContactItems.Type.UpdateContactItem;
    index: number;
    item: ResumeContactItem;
  }
}

export type ModifySections =
  | ModifySections.AddSection
  | ModifySections.MoveSection
  | ModifySections.RemoveSection
  | ModifySections.UpdateSection;

export namespace ModifySections {
  export enum Type {
    AddSection = "me.kthomas.resume.ModifySections.AddSection",
    MoveSection = "me.kthomas.resume.ModifySections.MoveSection",
    RemoveSection = "me.kthomas.resume.ModifySections.RemoveSection",
    UpdateSection = "me.kthomas.resume.ModifySections.UpdateSection",
  }
  
  export interface AddSection {
    type: ModifySections.Type.AddSection;
    section: ResumeSection;
  }
  
  export interface MoveSection {
    type: ModifySections.Type.MoveSection;
    oldIndex: number;
    newIndex: number;
  }
  
  export interface RemoveSection {
    type: ModifySections.Type.RemoveSection;
    index: number;
  }
  
  export interface UpdateSection {
    type: ModifySections.Type.UpdateSection;
    index: number;
    updater: SectionUpdater;
  }
}

export type SettingsUpdater =
  | SettingsUpdater.UpdateBaseFontSize;

export namespace SettingsUpdater {
  export enum Type {
    UpdateBaseFontSize = "me.kthomas.resume.SettingsUpdater.UpdateBaseFontSize",
  }
  
  export interface UpdateBaseFontSize {
    type: SettingsUpdater.Type.UpdateBaseFontSize;
    size: number;
  }
}

export interface ResumeContactItem {
  icon: ResumeContactIcon;
  value: string;
}

export interface ResumeSection {
  title: string;
  contentItems: ResumeSectionItem[];
}

export type SectionUpdater =
  | SectionUpdater.AddContent
  | SectionUpdater.MoveContent
  | SectionUpdater.RemoveContent
  | SectionUpdater.UpdateContent
  | SectionUpdater.UpdateTitle;

export namespace SectionUpdater {
  export enum Type {
    AddContent = "me.kthomas.resume.SectionUpdater.AddContent",
    MoveContent = "me.kthomas.resume.SectionUpdater.MoveContent",
    RemoveContent = "me.kthomas.resume.SectionUpdater.RemoveContent",
    UpdateContent = "me.kthomas.resume.SectionUpdater.UpdateContent",
    UpdateTitle = "me.kthomas.resume.SectionUpdater.UpdateTitle",
  }
  
  export interface AddContent {
    type: SectionUpdater.Type.AddContent;
    content: ResumeSectionItem;
  }
  
  export interface MoveContent {
    type: SectionUpdater.Type.MoveContent;
    oldIndex: number;
    newIndex: number;
  }
  
  export interface RemoveContent {
    type: SectionUpdater.Type.RemoveContent;
    index: number;
  }
  
  export interface UpdateContent {
    type: SectionUpdater.Type.UpdateContent;
    index: number;
    updater: SectionItemUpdater;
  }
  
  export interface UpdateTitle {
    type: SectionUpdater.Type.UpdateTitle;
    newTitle: string;
  }
}

export enum ResumeContactIcon {
  Email = "Email",
  Phone = "Phone",
  Website = "Website",
  Github = "Github",
  LinkedIn = "LinkedIn",
}

export interface ResumeSectionItem {
  leftLabel: string | null;
  centerLabel: string | null;
  rightLabel: string | null;
  contentItems: ResumeText[];
}

export type SectionItemUpdater =
  | SectionItemUpdater.AddContent
  | SectionItemUpdater.MoveContent
  | SectionItemUpdater.RemoveContent
  | SectionItemUpdater.UpdateLabels
  | SectionItemUpdater.UpdateText;

export namespace SectionItemUpdater {
  export enum Type {
    AddContent = "me.kthomas.resume.SectionItemUpdater.AddContent",
    MoveContent = "me.kthomas.resume.SectionItemUpdater.MoveContent",
    RemoveContent = "me.kthomas.resume.SectionItemUpdater.RemoveContent",
    UpdateLabels = "me.kthomas.resume.SectionItemUpdater.UpdateLabels",
    UpdateText = "me.kthomas.resume.SectionItemUpdater.UpdateText",
  }
  
  export interface AddContent {
    type: SectionItemUpdater.Type.AddContent;
    content: ResumeText;
  }
  
  export interface MoveContent {
    type: SectionItemUpdater.Type.MoveContent;
    oldIndex: number;
    newIndex: number;
  }
  
  export interface RemoveContent {
    type: SectionItemUpdater.Type.RemoveContent;
    index: number;
  }
  
  export interface UpdateLabels {
    type: SectionItemUpdater.Type.UpdateLabels;
    leftLabel?: string | null;
    centerLabel?: string | null;
    rightLabel?: string | null;
    useNull?: boolean;
  }
  
  export interface UpdateText {
    type: SectionItemUpdater.Type.UpdateText;
    index: number;
    updater: ResumeTextUpdater;
  }
}

export type ResumeText =
  | ResumeText.BulletPoints
  | ResumeText.Columns
  | ResumeText.Paragraph;

export namespace ResumeText {
  export enum Type {
    BulletPoints = "me.kthomas.resume.ResumeText.BulletPoints",
    Columns = "me.kthomas.resume.ResumeText.Columns",
    Paragraph = "me.kthomas.resume.ResumeText.Paragraph",
  }
  
  export interface BulletPoints {
    type: ResumeText.Type.BulletPoints;
    items: string[];
    columns: number;
  }
  
  export interface Columns {
    type: ResumeText.Type.Columns;
    items: ColumnItem[];
  }
  
  export interface Paragraph {
    type: ResumeText.Type.Paragraph;
    text: string;
  }
}

export type ResumeTextUpdater =
  | ResumeTextUpdater.AddBulletPoint
  | ResumeTextUpdater.RemoveBulletPoint
  | ResumeTextUpdater.UpdateBulletPoints
  | ResumeTextUpdater.UpdateColumnItem
  | ResumeTextUpdater.UpdateColumns
  | ResumeTextUpdater.UpdateParagraph;

export namespace ResumeTextUpdater {
  export enum Type {
    AddBulletPoint = "me.kthomas.resume.ResumeTextUpdater.AddBulletPoint",
    RemoveBulletPoint = "me.kthomas.resume.ResumeTextUpdater.RemoveBulletPoint",
    UpdateBulletPoints = "me.kthomas.resume.ResumeTextUpdater.UpdateBulletPoints",
    UpdateColumnItem = "me.kthomas.resume.ResumeTextUpdater.UpdateColumnItem",
    UpdateColumns = "me.kthomas.resume.ResumeTextUpdater.UpdateColumns",
    UpdateParagraph = "me.kthomas.resume.ResumeTextUpdater.UpdateParagraph",
  }
  
  export interface AddBulletPoint {
    type: ResumeTextUpdater.Type.AddBulletPoint;
    item: string;
  }
  
  export interface RemoveBulletPoint {
    type: ResumeTextUpdater.Type.RemoveBulletPoint;
    index: number;
  }
  
  export interface UpdateBulletPoints {
    type: ResumeTextUpdater.Type.UpdateBulletPoints;
    items?: string[] | null;
    columns?: number | null;
  }
  
  export interface UpdateColumnItem {
    type: ResumeTextUpdater.Type.UpdateColumnItem;
    index: number;
    updater: ColumnItemUpdater;
  }
  
  export interface UpdateColumns {
    type: ResumeTextUpdater.Type.UpdateColumns;
    items: ColumnItem[];
  }
  
  export interface UpdateParagraph {
    type: ResumeTextUpdater.Type.UpdateParagraph;
    text: string;
  }
}
// ...
export interface ColumnItem {
  label: string;
  items: string[];
}